msdocs-python-flask-webapp-quickstart/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.0.3
azure-core==1.1.1
azure-cosmos==4.1.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
six==1.13.0
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3
